+++ services/liaradb/collection/tenant/tenant.go
@@ -78,8 +78,7 @@ func (o *Tenant) getItem(ctx context.Context, tn tablename.TableName, rid link.R
 		return nil, page.ErrNotPage
 	}
 
-	// TODO: Fix this type
-	d, ok := n.Child(int16(rid.Position()))
+	d, ok := n.Child(rid.Position())
 	if !ok {
 		return nil, btree.ErrNotFound
 	}
